priorities router

+++ routers/priorities_router/crud.py
@@ -1,17 +1,8 @@
-from sqlalchemy.orm import Session
 from fastapi import Depends, HTTPException
-from sqlalchemy import update, and_, desc
-# import sys
-
-from main import get_db
-
-import utils
-
-import sys
-
+from sqlalchemy.orm import Session
+from sqlalchemy import exc, desc
 from . import models, schemas
-
-from sqlalchemy.exc import IntegrityError, SQLAlchemyError
+import sys
 
 async def create_priority(payload: schemas.PriorityCreate, db: Session):
     try:
@@ -20,20 +11,26 @@ async def create_priority(payload: schemas.PriorityCreate, db: Session):
         db.commit()
         db.refresh(new_priority)
         return new_priority
-    except IntegrityError as e:
+    except exc.IntegrityError as e:
         db.rollback()
-        raise HTTPException(status_code=400, detail="priority value should be unique")
+        print("{}".format(sys.exc_info()))
+        raise HTTPException(status_code=409)
     except:
         db.rollback()
+        print("{}".format(sys.exc_info()))
+        raise HTTPException(status_code=500)
 
 async def delete_priority(id: int, db: Session):
     try:
         priority = await read_priority_by_id(id, db)
-        db.delete(priority)
+        if priority:
+            db.delete(priority)
         db.commit()
         return True
     except:
         db.rollback()
+        print("{}".format(sys.exc_info()))
+        raise HTTPException(status_code=500)
 
 async def read_priority_by_id(id: int, db: Session):
     return db.query(models.Priorities).filter(models.Priorities.id == id).first()
@@ -48,18 +45,20 @@ async def read_priorities(skip: int, limit: int, search:str, value:str, db: Sess
     return base.order_by(desc(models.Priorities.priority)).offset(skip).limit(limit).all()
 
 async def update_priority(id: int, payload: schemas.PriorityUpdate, db: Session):
+    priority = await read_priority_by_id(id, db)
+    if not priority:
+        raise HTTPException(status_code=404)
     try:
-        priority = await read_priority_by_id(id, db)
-        if not priority:
-            raise HTTPException(status_code=404)
-        data = db.query(models.Priorities).filter(models.Priorities.id == id).update(payload.dict(exclude_unset=True))
+        updated = db.query(models.Priorities).filter(models.Priorities.id == id).update(payload.dict(exclude_unset=True))
         db.commit()
-        return bool(data)
-
-    except IntegrityError as e:
+        if bool(updated):
+            return await read_priority_by_id(id, db)
+    except exc.IntegrityError:
         db.rollback()
-        raise HTTPException(status_code=400, detail="priority value should be unique")
-
+        print("{}".format(sys.exc_info()))
+        raise HTTPException(status_code=409)
     except:
        db.rollback() 
+       print("{}".format(sys.exc_info()))
+       raise HTTPException(status_code=500)
